{"schema":"decodex/commit/1","summary":"Modularize diagnostics docs and recovery policy boundaries","authority":"manual"}
