+++ apps/decodex/src/orchestrator/execution_failure/zero_evidence.rs
@@ -1,65 +1,15 @@
-use std::{
-	env,
-	error::Error,
-	fmt::{Display, Formatter},
+mod context;
+mod diagnostic;
+mod error;
+mod record;
+
+pub(crate) use self::{
+	diagnostic::truncate_private_diagnostic_text, error::AppServerZeroEvidenceStartFailure,
 };
 
 use crate::orchestrator::execution_failure::{
 	self, IssueRunPlan, Report, ServiceConfig, StateStore,
 };
-
-#[derive(Debug)]
-pub(crate) struct AppServerZeroEvidenceStartFailure {
-	issue_identifier: String,
-	run_id: String,
-}
-impl AppServerZeroEvidenceStartFailure {
-	pub(crate) fn new(issue_identifier: String, run_id: String) -> Self {
-		Self { issue_identifier, run_id }
-	}
-
-	pub(crate) fn error_class(&self) -> &'static str {
-		"app_server_zero_evidence_start_failed"
-	}
-
-	pub(crate) fn terminal_next_action(&self, recovery_gate: &str) -> String {
-		format!(
-			"inspect local app-server startup logs and Decodex account/runtime state for run `{}`, verify `decodex probe stdio://`, restart `decodex serve` if needed, {recovery_gate}",
-			self.run_id
-		)
-	}
-
-	pub(crate) fn retry_next_action(&self) -> String {
-		format!(
-			"restart the app-server and retry automatically for run `{}`; inspect private startup diagnostics if the retry budget exhausts",
-			self.run_id
-		)
-	}
-}
-
-impl Display for AppServerZeroEvidenceStartFailure {
-	fn fmt(&self, formatter: &mut Formatter<'_>) -> std::fmt::Result {
-		write!(
-			formatter,
-			"App-server run `{}` for issue `{}` failed before Decodex recorded a thread, turn, protocol event, or private execution event.",
-			self.run_id, self.issue_identifier
-		)
-	}
-}
-
-impl Error for AppServerZeroEvidenceStartFailure {}
-
-pub(super) struct ZeroEvidenceAppServerStartFailureContext {
-	protocol_event_count: i64,
-	private_event_count: usize,
-	thread_recorded: bool,
-	turn_recorded: bool,
-}
-
-pub(super) struct ZeroEvidenceAppServerStartFailureDiagnostic {
-	source_error_summary: String,
-	source_error_chain: Vec<String>,
-}
 
 pub(crate) fn promote_zero_evidence_app_server_start_failure(
 	project: &ServiceConfig,
@@ -75,11 +25,11 @@ pub(crate) fn promote_zero_evidence_app_server_start_failure(
 		return error;
 	}
 
-	match zero_evidence_app_server_start_failure_context(project, state_store, issue_run) {
+	match context::zero_evidence_app_server_start_failure_context(project, state_store, issue_run) {
 		Ok(Some(context)) => {
-			let diagnostic = zero_evidence_app_server_start_failure_diagnostic(&error);
+			let diagnostic = diagnostic::zero_evidence_app_server_start_failure_diagnostic(&error);
 
-			if let Err(record_error) = record_zero_evidence_app_server_start_failure(
+			if let Err(record_error) = record::record_zero_evidence_app_server_start_failure(
 				project,
 				state_store,
 				issue_run,
@@ -118,127 +68,4 @@ pub(crate) fn promote_zero_evidence_app_server_start_failure(
 			error
 		},
 	}
-}
-
-pub(crate) fn truncate_private_diagnostic_text(text: &str) -> String {
-	const MAX_PRIVATE_DIAGNOSTIC_TEXT_CHARS: usize = 2_000;
-
-	if text.chars().count() <= MAX_PRIVATE_DIAGNOSTIC_TEXT_CHARS {
-		return text.to_owned();
-	}
-
-	let mut truncated = text.chars().take(MAX_PRIVATE_DIAGNOSTIC_TEXT_CHARS).collect::<String>();
-
-	truncated.push_str("...<truncated>");
-
-	truncated
-}
-
-fn zero_evidence_app_server_start_failure_context(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	issue_run: &IssueRunPlan,
-) -> crate::orchestrator::execution_failure::Result<Option<ZeroEvidenceAppServerStartFailureContext>>
-{
-	let protocol_event_count = state_store.event_count(&issue_run.run_id)?;
-	let private_event_count = state_store
-		.list_private_execution_events(
-			project.service_id(),
-			&issue_run.issue.id,
-			&issue_run.run_id,
-			issue_run.attempt_number,
-		)?
-		.len();
-	let run_attempt = state_store.run_attempt(&issue_run.run_id)?;
-	let thread_recorded = run_attempt.as_ref().and_then(|attempt| attempt.thread_id()).is_some();
-	let turn_recorded = run_attempt.as_ref().and_then(|attempt| attempt.turn_id()).is_some();
-
-	if protocol_event_count == 0 && private_event_count == 0 && !thread_recorded && !turn_recorded {
-		Ok(Some(ZeroEvidenceAppServerStartFailureContext {
-			protocol_event_count,
-			private_event_count,
-			thread_recorded,
-			turn_recorded,
-		}))
-	} else {
-		Ok(None)
-	}
-}
-
-fn record_zero_evidence_app_server_start_failure(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	issue_run: &IssueRunPlan,
-	context: &ZeroEvidenceAppServerStartFailureContext,
-	diagnostic: &ZeroEvidenceAppServerStartFailureDiagnostic,
-) -> crate::orchestrator::execution_failure::Result<()> {
-	state_store
-		.append_private_execution_event(
-			project.service_id(),
-			&issue_run.issue.id,
-			&issue_run.run_id,
-			issue_run.attempt_number,
-			"app_server_zero_evidence_start_failure",
-			serde_json::json!({
-				"error_class": "app_server_zero_evidence_start_failed",
-				"summary": "App-server dispatch failed before Decodex recorded a thread, turn, protocol event, or private execution event.",
-				"issue_identifier": issue_run.issue.identifier.as_str(),
-				"attempt_number": issue_run.attempt_number,
-				"branch": issue_run.worktree.branch_name.as_str(),
-				"worktree_path": issue_run.worktree.path.display().to_string(),
-				"protocol_event_count": context.protocol_event_count,
-				"private_event_count": context.private_event_count,
-				"thread_recorded": context.thread_recorded,
-				"turn_recorded": context.turn_recorded,
-				"source_error_summary": diagnostic.source_error_summary.as_str(),
-				"source_error_chain": &diagnostic.source_error_chain,
-			}),
-		)
-		.map(|_| ())
-}
-
-fn zero_evidence_app_server_start_failure_diagnostic(
-	error: &Report,
-) -> ZeroEvidenceAppServerStartFailureDiagnostic {
-	let source_error_chain = error
-		.chain()
-		.map(|cause| sanitize_private_diagnostic_text(&cause.to_string()))
-		.collect::<Vec<_>>();
-	let source_error_summary = source_error_chain
-		.first()
-		.cloned()
-		.unwrap_or_else(|| String::from("unknown app-server startup failure"));
-
-	ZeroEvidenceAppServerStartFailureDiagnostic { source_error_summary, source_error_chain }
-}
-
-fn sanitize_private_diagnostic_text(text: &str) -> String {
-	let mut sanitized = text.to_owned();
-
-	for (name, value) in env::vars() {
-		if !diagnostic_env_var_name_is_sensitive(&name) || value.len() < 6 {
-			continue;
-		}
-
-		let replacement = format!("<redacted env:{name}>");
-
-		sanitized = sanitized.replace(&value, &replacement);
-	}
-
-	truncate_private_diagnostic_text(&sanitized)
-}
-
-fn diagnostic_env_var_name_is_sensitive(name: &str) -> bool {
-	let normalized = name.to_ascii_lowercase();
-
-	normalized.contains("token")
-		|| normalized.contains("secret")
-		|| normalized.contains("password")
-		|| normalized.contains("credential")
-		|| normalized.contains("api_key")
-		|| normalized.contains("apikey")
-		|| normalized.ends_with("_pat")
-		|| normalized.starts_with("pat_")
-		|| normalized.contains("_pat_")
-		|| normalized.contains("auth")
 }
