+++ apps/decodex/src/orchestrator/execution_failure/zero_evidence/context.rs
@@ -1,0 +1,38 @@
+use crate::orchestrator::execution_failure::{IssueRunPlan, Result, ServiceConfig, StateStore};
+
+pub(super) struct ZeroEvidenceAppServerStartFailureContext {
+	pub(super) protocol_event_count: i64,
+	pub(super) private_event_count: usize,
+	pub(super) thread_recorded: bool,
+	pub(super) turn_recorded: bool,
+}
+
+pub(super) fn zero_evidence_app_server_start_failure_context(
+	project: &ServiceConfig,
+	state_store: &StateStore,
+	issue_run: &IssueRunPlan,
+) -> Result<Option<ZeroEvidenceAppServerStartFailureContext>> {
+	let protocol_event_count = state_store.event_count(&issue_run.run_id)?;
+	let private_event_count = state_store
+		.list_private_execution_events(
+			project.service_id(),
+			&issue_run.issue.id,
+			&issue_run.run_id,
+			issue_run.attempt_number,
+		)?
+		.len();
+	let run_attempt = state_store.run_attempt(&issue_run.run_id)?;
+	let thread_recorded = run_attempt.as_ref().and_then(|attempt| attempt.thread_id()).is_some();
+	let turn_recorded = run_attempt.as_ref().and_then(|attempt| attempt.turn_id()).is_some();
+
+	if protocol_event_count == 0 && private_event_count == 0 && !thread_recorded && !turn_recorded {
+		Ok(Some(ZeroEvidenceAppServerStartFailureContext {
+			protocol_event_count,
+			private_event_count,
+			thread_recorded,
+			turn_recorded,
+		}))
+	} else {
+		Ok(None)
+	}
+}
