+++ apps/decodex/src/orchestrator/execution_failure/zero_evidence/record.rs
@@ -1,0 +1,39 @@
+use crate::orchestrator::execution_failure::{
+	IssueRunPlan, Result, ServiceConfig, StateStore,
+	zero_evidence::{
+		context::ZeroEvidenceAppServerStartFailureContext,
+		diagnostic::ZeroEvidenceAppServerStartFailureDiagnostic,
+	},
+};
+
+pub(super) fn record_zero_evidence_app_server_start_failure(
+	project: &ServiceConfig,
+	state_store: &StateStore,
+	issue_run: &IssueRunPlan,
+	context: &ZeroEvidenceAppServerStartFailureContext,
+	diagnostic: &ZeroEvidenceAppServerStartFailureDiagnostic,
+) -> Result<()> {
+	state_store
+		.append_private_execution_event(
+			project.service_id(),
+			&issue_run.issue.id,
+			&issue_run.run_id,
+			issue_run.attempt_number,
+			"app_server_zero_evidence_start_failure",
+			serde_json::json!({
+				"error_class": "app_server_zero_evidence_start_failed",
+				"summary": "App-server dispatch failed before Decodex recorded a thread, turn, protocol event, or private execution event.",
+				"issue_identifier": issue_run.issue.identifier.as_str(),
+				"attempt_number": issue_run.attempt_number,
+				"branch": issue_run.worktree.branch_name.as_str(),
+				"worktree_path": issue_run.worktree.path.display().to_string(),
+				"protocol_event_count": context.protocol_event_count,
+				"private_event_count": context.private_event_count,
+				"thread_recorded": context.thread_recorded,
+				"turn_recorded": context.turn_recorded,
+				"source_error_summary": diagnostic.source_error_summary.as_str(),
+				"source_error_chain": &diagnostic.source_error_chain,
+			}),
+		)
+		.map(|_| ())
+}
