+++ apps/decodex/src/orchestrator/status/review_orchestration/classification.rs
@@ -1,242 +1,9 @@
-use crate::{
-	orchestrator::status::{
-		self, EXTERNAL_REVIEW_MERGE_VISIBILITY_TIMEOUT_SECS, ExternalReviewRequestCiGate,
-		PostReviewLaneClassification, PostReviewLaneDecision, PostReviewLaneSnapshot,
-		PostReviewOrchestrationStatus, PullRequestReviewState, ReviewOrchestrationMarker,
-		ReviewOrchestrationPhase, WorkflowDocument,
-	},
-	prelude::{Result, eyre},
+mod external;
+mod non_github;
+mod pre;
+
+pub(crate) use self::{
+	external::apply_review_orchestration_phase_classification,
+	non_github::apply_non_github_review_post_review_classification,
+	pre::apply_pre_orchestration_post_review_classification,
 };
-
-pub(crate) fn apply_pre_orchestration_post_review_classification(
-	snapshot: &PostReviewLaneSnapshot,
-	workflow: &WorkflowDocument,
-	review_state: &PullRequestReviewState,
-	classification: &mut PostReviewLaneClassification,
-) -> bool {
-	if review_state.state == "MERGED" {
-		classification.decision = PostReviewLaneDecision::Continue;
-		classification.reason = String::from("pull_request_merged_closeout_pending");
-
-		return true;
-	}
-	if snapshot.issue.state.name == workflow.frontmatter().tracker().resolved_completed_state() {
-		*classification = status::blocked_post_review_lane_from_state(
-			review_state,
-			"issue_completed_before_pull_request_merged",
-		);
-
-		return true;
-	}
-	if review_state.state != "OPEN" {
-		*classification =
-			status::blocked_post_review_lane_from_state(review_state, "pull_request_not_open");
-
-		return true;
-	}
-	if review_state.is_draft {
-		*classification =
-			status::blocked_post_review_lane_from_state(review_state, "pull_request_is_draft");
-
-		return true;
-	}
-	if review_state.unresolved_review_threads > 0 {
-		classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-		classification.reason = String::from("unresolved_review_threads");
-
-		return true;
-	}
-	if matches!(review_state.review_decision.as_deref(), Some("CHANGES_REQUESTED")) {
-		classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-		classification.reason = String::from("review_changes_requested");
-
-		return true;
-	}
-	if status::failed_checks_require_repair(
-		review_state.status_check_rollup_state.as_deref(),
-		&review_state.merge_state_status,
-	) {
-		classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-		classification.reason = String::from("required_checks_failed");
-
-		return true;
-	}
-
-	if let Some(reason) = status::merge_state_requires_review_repair(
-		&review_state.mergeable,
-		&review_state.merge_state_status,
-	) {
-		classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-		classification.reason = String::from(reason);
-
-		return true;
-	}
-
-	false
-}
-
-pub(crate) fn apply_non_github_review_post_review_classification(
-	classification: &mut PostReviewLaneClassification,
-	review_state: &PullRequestReviewState,
-	orchestration_marker: Option<&ReviewOrchestrationMarker>,
-	now_unix_epoch: i64,
-) -> Result<()> {
-	if let Some(orchestration_marker) = orchestration_marker {
-		let phase =
-			ReviewOrchestrationPhase::parse(orchestration_marker.phase()).map_err(|error| {
-				eyre::eyre!("Failed to parse retained review orchestration phase: {error}")
-			})?;
-
-		if phase == ReviewOrchestrationPhase::WaitingForMerge {
-			if let Some(auto_merge_enabled_at) =
-				orchestration_marker.auto_merge_enabled_at_unix_epoch()
-				&& now_unix_epoch - auto_merge_enabled_at
-					> EXTERNAL_REVIEW_MERGE_VISIBILITY_TIMEOUT_SECS
-			{
-				*classification = status::blocked_post_review_lane_from_state(
-					review_state,
-					"non_github_review_merge_visibility_timeout",
-				);
-			} else {
-				classification.reason = String::from("non_github_review_waiting_for_merge");
-			}
-
-			return Ok(());
-		}
-		if phase == ReviewOrchestrationPhase::RepairRequired {
-			classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-			classification.reason =
-				if status::review_state_landing_requires_agent_fallback(review_state) {
-					String::from("retained_landing_agent_fallback_required")
-				} else {
-					String::from("non_github_review_repair_required")
-				};
-
-			return Ok(());
-		}
-	}
-
-	if status::review_state_clean_path_landing_gates_satisfied(review_state) {
-		classification.decision = PostReviewLaneDecision::ReadyToLand;
-		classification.reason = String::from("non_github_review_ready_to_land");
-	} else if status::review_state_landing_requires_agent_fallback(review_state) {
-		classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-		classification.reason = String::from("retained_landing_agent_fallback_required");
-	} else {
-		classification.reason = String::from("non_github_review_waiting_gates");
-	}
-
-	Ok(())
-}
-
-pub(crate) fn apply_review_orchestration_phase_classification(
-	classification: &mut PostReviewLaneClassification,
-	review_state: &PullRequestReviewState,
-	orchestration_marker: &ReviewOrchestrationMarker,
-	orchestration_status: &PostReviewOrchestrationStatus,
-	now_unix_epoch: i64,
-) {
-	match orchestration_status.phase {
-		ReviewOrchestrationPhase::RequestPending => {
-			match status::external_review_request_ci_gate(review_state) {
-				ExternalReviewRequestCiGate::Ready => {
-					classification.reason = String::from("external_review_request_pending");
-				},
-				ExternalReviewRequestCiGate::WaitForGreenChecks => {
-					classification.reason =
-						String::from("external_review_request_waiting_for_green_checks");
-				},
-				ExternalReviewRequestCiGate::RepairRequired => {
-					classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-					classification.reason =
-						String::from("external_review_request_ci_red_repair_required");
-				},
-			}
-		},
-		ReviewOrchestrationPhase::WaitingForAck =>
-			if orchestration_status.request_acknowledged {
-				classification.reason = String::from("external_review_result_pending");
-			} else if status::request_ack_timed_out(orchestration_marker, now_unix_epoch) {
-				*classification = status::blocked_post_review_lane_from_state(
-					review_state,
-					"external_review_ack_timeout",
-				);
-			} else {
-				classification.reason = String::from("external_review_ack_pending");
-			},
-		ReviewOrchestrationPhase::WaitingForResult => {
-			if !orchestration_status.request_acknowledged {
-				classification.reason = String::from("external_review_ack_pending");
-			} else if status::external_review_has_actionable_feedback(
-				review_state,
-				orchestration_marker,
-			) {
-				classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-				classification.reason = String::from("external_review_feedback_pending_repair");
-			} else if orchestration_status.strict_pass
-				&& orchestration_status.clean_path_landing_gates_satisfied
-			{
-				classification.decision = PostReviewLaneDecision::ReadyToLand;
-				classification.reason = String::from("external_review_passed_strict");
-			} else if orchestration_status.strict_pass
-				&& orchestration_status.landing_requires_agent_fallback
-			{
-				classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-				classification.reason = String::from("retained_landing_agent_fallback_required");
-			} else if orchestration_status.strict_pass {
-				classification.reason = String::from("external_review_passed_waiting_gates");
-			} else if orchestration_status.review_result_arrived {
-				*classification = status::blocked_post_review_lane_from_state(
-					review_state,
-					"external_review_pass_signal_missing",
-				);
-			} else {
-				classification.reason = String::from("external_review_result_pending");
-			}
-		},
-		ReviewOrchestrationPhase::RepairRequired => {
-			classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-			classification.reason = if orchestration_status.landing_requires_agent_fallback {
-				String::from("retained_landing_agent_fallback_required")
-			} else {
-				String::from("external_review_feedback_pending_repair")
-			};
-		},
-		ReviewOrchestrationPhase::PassWaitingForGates =>
-			if status::external_review_has_actionable_feedback(review_state, orchestration_marker) {
-				classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-				classification.reason = String::from("external_review_feedback_pending_repair");
-			} else if orchestration_status.strict_pass
-				&& orchestration_status.clean_path_landing_gates_satisfied
-			{
-				classification.decision = PostReviewLaneDecision::ReadyToLand;
-				classification.reason = String::from("external_review_passed_strict");
-			} else if orchestration_status.strict_pass
-				&& orchestration_status.landing_requires_agent_fallback
-			{
-				classification.decision = PostReviewLaneDecision::NeedsReviewRepair;
-				classification.reason = String::from("retained_landing_agent_fallback_required");
-			} else if orchestration_status.strict_pass {
-				classification.reason = String::from("external_review_passed_waiting_gates");
-			} else {
-				*classification = status::blocked_post_review_lane_from_state(
-					review_state,
-					"external_review_pass_signal_missing",
-				);
-			},
-		ReviewOrchestrationPhase::WaitingForMerge => {
-			if let Some(auto_merge_enabled_at) =
-				orchestration_marker.auto_merge_enabled_at_unix_epoch()
-				&& now_unix_epoch - auto_merge_enabled_at
-					> EXTERNAL_REVIEW_MERGE_VISIBILITY_TIMEOUT_SECS
-			{
-				*classification = status::blocked_post_review_lane_from_state(
-					review_state,
-					"external_review_merge_visibility_timeout",
-				);
-			} else {
-				classification.reason = String::from("external_review_waiting_for_merge");
-			}
-		},
-	}
-}
