+++ apps/decodex/src/recovery/review_handoff_policy/adopt.rs
@@ -1,0 +1,95 @@
+use crate::{
+	prelude::{Result, eyre},
+	pull_request::{self, LandingGateMode, PullRequestLandingGateView, PullRequestLandingState},
+	recovery,
+};
+
+pub(in crate::recovery) fn validate_adopt_landing_state(
+	landing_state: &PullRequestLandingState,
+) -> Result<()> {
+	let pr_url = recovery::landing_url(landing_state);
+	let gate_view = landing_state.gate_view();
+	let decision = pull_request::classify_landing_gate(gate_view, LandingGateMode::Adopt);
+
+	match decision {
+		pull_request::LandingGateDecision::Satisfied => Ok(()),
+		decision => adopt_landing_gate_error(decision, gate_view, pr_url),
+	}
+}
+
+fn adopt_landing_gate_error(
+	decision: pull_request::LandingGateDecision,
+	gate_view: PullRequestLandingGateView<'_>,
+	pr_url: &str,
+) -> Result<()> {
+	match decision {
+		pull_request::LandingGateDecision::Satisfied => Ok(()),
+		pull_request::LandingGateDecision::CloseoutOnly
+		| pull_request::LandingGateDecision::Block("pull_request_not_open") => {
+			eyre::bail!("Pull request `{pr_url}` is `{}`; adopt requires `OPEN`.", gate_view.state)
+		},
+		pull_request::LandingGateDecision::Block("pull_request_is_draft") => {
+			eyre::bail!("Pull request `{pr_url}` is still draft.")
+		},
+		pull_request::LandingGateDecision::Wait("pending_review_requests") => {
+			eyre::bail!(
+				"Pull request `{pr_url}` still has {} pending review request(s).",
+				gate_view.pending_review_requests
+			)
+		},
+		pull_request::LandingGateDecision::Repair("unresolved_review_threads") => {
+			eyre::bail!(
+				"Pull request `{pr_url}` still has {} unresolved review thread(s).",
+				gate_view.unresolved_review_threads
+			)
+		},
+		pull_request::LandingGateDecision::Repair("review_changes_requested") => {
+			eyre::bail!("Pull request `{pr_url}` still has active change requests.")
+		},
+		pull_request::LandingGateDecision::Repair(reason)
+			if matches!(
+				reason,
+				"pull_request_merge_conflict" | "pull_request_branch_behind_base"
+			) =>
+		{
+			eyre::bail!("Pull request `{pr_url}` requires review repair: {reason}.")
+		},
+		pull_request::LandingGateDecision::Repair("required_checks_failed") => {
+			eyre::bail!("Pull request `{pr_url}` has failed required checks that need repair.")
+		},
+		pull_request::LandingGateDecision::Wait("checks_waiting") => {
+			let check_state = gate_view.status_check_rollup_state.unwrap_or("unknown");
+
+			eyre::bail!(
+				"Pull request `{pr_url}` is still waiting on checks: statusCheckRollup=`{check_state}`."
+			)
+		},
+		pull_request::LandingGateDecision::Wait("mergeability_unknown") => {
+			eyre::bail!("Pull request `{pr_url}` mergeability is still unknown.")
+		},
+		pull_request::LandingGateDecision::Block("merge_state_not_ready") => {
+			eyre::bail!(
+				"Pull request `{pr_url}` is not ready to adopt: mergeStateStatus=`{}`.",
+				gate_view.merge_state_status
+			)
+		},
+		pull_request::LandingGateDecision::Block("not_mergeable") => {
+			eyre::bail!(
+				"Pull request `{pr_url}` is not mergeable: mergeable=`{}`.",
+				gate_view.mergeable
+			)
+		},
+		pull_request::LandingGateDecision::Wait("checks_non_green") => {
+			let check_state = gate_view.status_check_rollup_state.unwrap_or("unknown");
+
+			eyre::bail!(
+				"Pull request `{pr_url}` still has non-green checks: statusCheckRollup=`{check_state}`."
+			)
+		},
+		pull_request::LandingGateDecision::Wait(reason)
+		| pull_request::LandingGateDecision::Repair(reason)
+		| pull_request::LandingGateDecision::Block(reason) => {
+			eyre::bail!("Pull request `{pr_url}` is not ready to adopt: {reason}.")
+		},
+	}
+}
